// ui-catalogue/src/components/molecules/HistoryPeru/types.tsx
interface Ii18nHistoryPeru {
  'HistoryPeru:title': string
  'HistoryPeru:nroEntidades': string
  'HistoryPeru:nor': string
  'HistoryPeru:cpp': string
  'HistoryPeru:def': string
  'HistoryPeru:dud': string
  'HistoryPeru:per': string
  'HistoryPeru:totalDeuda': string
}

export interface IHistoryPeruRequestForm {
  regulatedPeriodo1: string
  regulatedPeriodo2: string
  regulatedPeriodo3: string
  regulatedPeriodo4: string
  regulatedPeriodo5: string
  regulatedPeriodo6: string
  regulatednroEntidades6: string
  regulatedNroEntidades1: string
  regulatedNroEntidades2: string
  regulatedNroEntidades3: string
  regulatedNroEntidades4: string
  regulatedNroEntidades5: string
  regulatedNroEntidades6: string
  regulatedNor1: string
  regulatedNor2: string
  regulatedNor3: string
  regulatedNor4: string
  regulatedNor5: string
  regulatedNor6: string
  regulatedCpp1: string
  regulatedCpp2: string
  regulatedCpp3: string
  regulatedCpp4: string
  regulatedCpp5: string
  regulatedCpp6: string
  regulatedDef1: string
  regulatedDef2: string
  regulatedDef3: string
  regulatedDef4: string
  regulatedDef5: string
  regulatedDef6: string
  regulatedDud1: string
  regulatedDud2: string
  regulatedDud3: string
  regulatedDud4: string
  regulatedDud5: string
  regulatedDud6: string
  regulatedPer1: string
  regulatedPer2: string
  regulatedPer3: string
  regulatedPer4: string
  regulatedPer5: string
  regulatedPer6: string
  regulatedTotalDeuda1: string
  regulatedTotalDeuda2: string
  regulatedTotalDeuda3: string
  regulatedTotalDeuda4: string
  regulatedTotalDeuda5: string
  regulatedTotalDeuda6: string
  nonRegulatedPeriodo1: string
  nonRegulatedPeriodo2: string
  nonRegulatedPeriodo3: string
  nonRegulatedPeriodo4: string
  nonRegulatedPeriodo5: string
  nonRegulatedPeriodo6: string
  nonRegulatedNroEntidades1: string
  nonRegulatedNroEntidades2: string
  nonRegulatedNroEntidades3: string
  nonRegulatedNroEntidades4: string
  nonRegulatedNroEntidades5: string
  nonRegulatedNroEntidades6: string
  nonRegulatedNor1: string
  nonRegulatedNor2: string
  nonRegulatedNor3: string
  nonRegulatedNor4: string
  nonRegulatedNor5: string
  nonRegulatedNor6: string
  nonRegulatedCpp1: string
  nonRegulatedCpp2: string
  nonRegulatedCpp3: string
  nonRegulatedCpp4: string
  nonRegulatedCpp5: string
  nonRegulatedCpp6: string
  nonRegulatedDef1: string
  nonRegulatedDef2: string
  nonRegulatedDef3: string
  nonRegulatedDef4: string
  nonRegulatedDef5: string
  nonRegulatedDef6: string
  nonRegulatedDud1: string
  nonRegulatedDud2: string
  nonRegulatedDud3: string
  nonRegulatedDud4: string
  nonRegulatedDud5: string
  nonRegulatedDud6: string
  nonRegulatedPer1: string
  nonRegulatedPer2: string
  nonRegulatedPer3: string
  nonRegulatedPer4: string
  nonRegulatedPer5: string
  nonRegulatedPer6: string
  nonRegulatedTotalDeuda1: string
  nonRegulatedTotalDeuda2: string
  nonRegulatedTotalDeuda3: string
  nonRegulatedTotalDeuda4: string
  nonRegulatedTotalDeuda5: string
  nonRegulatedTotalDeuda6: string
}

export interface IHistoryPeru {
  i18n: Ii18nHistoryPeru
  data: IHistoryPeruRequestForm
}